twitter-service completed

+++ auth-service/src/main/java/com/twitter/auth/service/Model/RefreshToken.java
@@ -17,7 +17,7 @@ public class RefreshToken {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer tokenId;
+    private Long tokenId;
 
     @Column(unique = true)
     private String token;
